Improve UI fomr list and fixed details on edge cases

+++ app/src/main/java/com/paypal/redditop/models/SimplePost.kt
@@ -19,7 +19,7 @@ data class SimplePost(
     val category: String,
     val subreddit: String,
     val subredditUrl: String,
-    val created: Int,
+    val created: Long,
     val isOriginalContent: Boolean,
     val isNsfw: Boolean,
     val isVideo: Boolean,
